covergan_backend/requirements.txt
# Server dependencies
cherrypy==18.1.1
psutil
python-magic
pyyaml